refactored LoginView to reuse shared logic
